hubeau-client/hubeau-client/pyproject.toml
[tool.poetry]
name = "hubeau-client"
version = "0.1.0"
description = ""
authors = ["Guillaume Pumey <guillaume.plumey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.32.3"
pandas = "~1"
google-cloud = "^0.34.0"
google-cloud-bigquery = "^3.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
